Refactor semian ticket management into its own files
